requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.2.10
attrs==19.3.0
Django==3.0.8
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
django-filter==2.2.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytz==2020.1
packaging==20.3
pluggy==0.13.1
py==1.8.1
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
python-memcached==1.59
sqlparse==0.3.1
six==1.14.0
certifi==2020.4.5.1
chardet==3.0.4
flake8==3.8.1
idna==2.9
importlib-metadata==1.6.0
mccabe==0.6.1
more-itertools==8.2.0
requests==2.23.0
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0